Fix server tests
